# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/articles/{articleId}" : {
      "get" : {
        "tags" : [ "article-controller" ],
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "article-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateArticleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "article-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/jobs/{jobId}/restore" : {
      "post" : {
        "tags" : [ "job-controller" ],
        "operationId" : "restoreJob",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/restore" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "restore",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/reopen" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "reopen",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/reject" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "reject",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RejectArticleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        },
        "deprecated" : true
      }
    },
    "/api/v1/articles/{articleId}/publications" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "publish",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PublishArticleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/publication-batches" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "publishBatch",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PublishArticleBatchRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicationBatchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/english-translations" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "translate",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/EnglishTranslationRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnglishTranslationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/{articleId}/confirm" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "confirm",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/approve" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "approve",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        },
        "deprecated" : true
      }
    },
    "/api/v1/articles/website-generations" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "generateWebsiteArticle",
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateWebsiteArticleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/topic-generations" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "generateTopicArticle",
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTopicArticleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/articles/custom" : {
      "post" : {
        "tags" : [ "article-controller" ],
        "operationId" : "createCustom",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateCustomArticleRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/monitoring/summary" : {
      "get" : {
        "tags" : [ "monitoring-controller" ],
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "range",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MonitoringSnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{jobId}" : {
      "get" : {
        "tags" : [ "job-controller" ],
        "operationId" : "getJob",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "job-controller" ],
        "operationId" : "deleteJob",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
...
    "/api/v1/articles/{articleId}/versions" : {
      "get" : {
        "tags" : [ "article-controller" ],
        "operationId" : "versions",
        "parameters" : [ {
          "name" : "articleId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ArticleVersionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UpdateArticleRequest" : {
        "type" : "object",
        "properties" : {
          "expectedVersion" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "title" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "summary" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "markdown" : {
            "type" : "string",
            "maxLength" : 100000,
            "minLength" : 0
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 50,
              "minLength" : 0
            },
            "maxItems" : 15,
            "minItems" : 0
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 100,
              "minLength" : 0
            },
            "maxItems" : 30,
            "minItems" : 0
          },
          "titleEn" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "summaryEn" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "markdownEn" : {
            "type" : "string",
            "maxLength" : 100000,
            "minLength" : 0
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 50,
              "minLength" : 0
            },
            "maxItems" : 15,
            "minItems" : 0
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 100,
              "minLength" : 0
            },
            "maxItems" : 30,
            "minItems" : 0
          }
        },
        "required" : [ "markdown", "summary", "title" ]
      },
      "ArticleResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "projectId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sourceType" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "sourceTitle" : {
            "type" : "string"
          },
          "sourceDescription" : {
            "type" : "string"
          },
          "targetAudience" : {
            "type" : "string"
          },
          "articleType" : {
            "type" : "string"
          },
          "knowledgeLevel" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "markdown" : {
            "type" : "string"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "titleEn" : {
            "type" : "string"
          },
          "summaryEn" : {
            "type" : "string"
          },
          "markdownEn" : {
            "type" : "string"
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "language" : {
            "type" : "string"
          },
          "sourceRevision" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "currentVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "JobResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "type" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "attempt" : {
            "type" : "integer",
            "format" : "int32"
          },
          "maxAttempts" : {
            "type" : "integer",
            "format" : "int32"
          },
          "progressPercent" : {
            "type" : "integer",
            "format" : "int32"
          },
          "progressLabel" : {
            "type" : "string"
          },
          "progressDetail" : {
            "type" : "string"
          },
          "batchId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "scheduledAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "resultResourceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "resultResourceType" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "errorMessage" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "RejectArticleRequest" : {
        "type" : "object",
        "properties" : {
          "reason" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          }
        },
        "required" : [ "reason" ]
      },
      "PublishArticleRequest" : {
        "type" : "object",
        "properties" : {
          "channelAccountId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "canonicalUrl" : {
            "type" : "string",
            "maxLength" : 2048,
            "minLength" : 0
          },
          "scheduledAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "channelAccountId" ]
      },
      "PublishArticleBatchRequest" : {
        "type" : "object",
        "properties" : {
          "channelAccountIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "maxItems" : 20,
            "minItems" : 0
          },
          "canonicalUrl" : {
            "type" : "string",
            "maxLength" : 2048,
            "minLength" : 0
          },
          "scheduledAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        },
        "required" : [ "channelAccountIds" ]
      },
      "PublicationBatchResponse" : {
        "type" : "object",
        "properties" : {
          "jobCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "jobs" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/JobResponse"
            }
          }
        }
      },
      "EnglishTranslationRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "summary" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "markdown" : {
            "type" : "string",
            "maxLength" : 100000,
            "minLength" : 0
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 50,
              "minLength" : 0
            },
            "maxItems" : 15,
            "minItems" : 0
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "maxLength" : 100,
              "minLength" : 0
            },
            "maxItems" : 30,
            "minItems" : 0
          }
        }
      },
      "EnglishTranslationResponse" : {
        "type" : "object",
        "properties" : {
          "titleEn" : {
            "type" : "string"
          },
          "summaryEn" : {
            "type" : "string"
          },
          "markdownEn" : {
            "type" : "string"
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "CreateCustomArticleRequest" : {
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "summary" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "markdown" : {
            "type" : "string",
            "maxLength" : 100000,
            "minLength" : 0
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywords" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "titleEn" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "summaryEn" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "markdownEn" : {
            "type" : "string",
            "maxLength" : 100000,
            "minLength" : 0
          },
          "tagsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "keywordsEn" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "language" : {
            "type" : "string",
            "maxLength" : 20,
            "minLength" : 0
          }
        },
        "required" : [ "markdown", "title" ]
      },
...
    }
  }
}